Merged 'v3' into bettermod
Squashed commit of the following:

[V3 Say] Bump to version 1.4.6

[V3 Say] Update french translations

[V3 i18n] Include docstrings in translations

[V3 Travis] Include every Python version in matrix

[V3 Travis] Include python 3.7 in matrix

[V3 Travis] Use int instead of str for python language

[V3] Add Makefile
This includes the following things :
- reformat : format the code with black
- stylecheck : only check for the code style, do not write
- gettext : generate .pot files
- compile : make python executables
- deploy : make the info.json files, used by Travis
- docs : compile the docs into HTML files

Black reformat

Make the class inherit from redbot.core.commands.Cog
This is because of a braking change introduced in Red RC1.
Merge of pull request #32 by @ASSASSIN0831

[V3 Say] Fix permission errors with channel destination
This also increases the time before deleting a message explaining an error.
Merge of pull request #33 by @ASSASSIN0831

[Black] Do not reformat int