Warnsystem version 1.4 (#117)
I am merging this PR now, as big updates simply do not fit my schedule anymore. I will now do more frequent updates with less features. This also mean minor version bumps will not contain as many features as you had in versions 1.1 to 1.3

The liked issues that were not completed for this release will be kept open, and may be done in the future.
As an engineer student, free time is starting to become less present. If anyone wants to help maintaining the repo, you're welcome to do so and seek my help if needed.

* Bump version number

* Add end user data functions

* Whitelist for antispam (not tested yet)

* Toggle for checking edited messages with automod

Not tested yet
Resolves #111